#67 Convert s3 and sqs service into DI classes
